requirements.txt
absl-py==2.0.0
accelerate==0.24.0
aiohttp==3.8.6
aiosignal==1.3.1
albumentations==1.3.1
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
bezier==2023.7.28
Brotli @ file:///tmp/abs_ecyw11_7ze/croots/recipe/brotli-split_1659616059936/work
cachetools==5.3.2
certifi==2022.12.7
cffi @ file:///croot/cffi_1670423208954/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
comm==0.1.4
contourpy==1.1.1
cryptography @ file:///croot/cryptography_1694444244250/work
cycler==0.12.1
datasets==2.14.6
debugpy==1.8.0
decorator==5.1.1
-e git+https://github.com/huggingface/diffusers.git@c697c5ab57eb9bc98a38d8aba9e2fc8fb0b06b95#egg=diffusers
dill==0.3.7
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==2.0.0
filelock @ file:///croot/filelock_1672387128942/work
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.10.0
ftfy==6.1.1
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455533097/work
google-auth==2.23.3
google-auth-oauthlib==1.1.0
grpcio==1.59.0
huggingface-hub==0.17.3
idna @ file:///croot/idna_1666125576474/work
imageio==2.31.6
importlib-metadata==6.8.0
ipykernel==6.26.0
ipython==8.16.1
jedi==0.19.1
Jinja2 @ file:///croot/jinja2_1666908132255/work
joblib==1.3.2
jupyter_client==8.5.0
jupyter_core==5.4.0
kiwisolver==1.4.5
lazy_loader==0.3
Markdown==3.5
MarkupSafe @ file:///opt/conda/conda-bld/markupsafe_1654597864307/work
matplotlib==3.8.0
matplotlib-inline==0.1.6
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mpmath @ file:///croot/mpmath_1690848262763/work
multidict==6.0.4
multiprocess==0.70.15
nest-asyncio==1.5.8
networkx==3.0
numpy @ file:///croot/numpy_and_numpy_base_1695830428084/work/dist/numpy-1.26.0-cp310-cp310-linux_x86_64.whl#sha256=fc2732718bc9e06a7b702492cb4f5afffe9671083930452d894377bf563464a3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
openpyxl==3.1.2
packaging==23.2
pandas==2.1.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow @ file:///croot/pillow_1696580024257/work
platformdirs==3.11.0
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments==2.16.1
pyOpenSSL @ file:///croot/pyopenssl_1690223430423/work
pyparsing==3.1.1
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
qudida==0.0.4
regex==2023.10.3
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.0
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.13.2
six==1.16.0
stack-data==0.6.3
sympy @ file:///croot/sympy_1668202399572/work
tensorboard==2.15.0
tensorboard-data-server==0.7.2
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2023.9.26
timm==0.9.8
tokenizers==0.14.1
torch==2.1.0+cu118
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.12.0
transformers==4.34.1
triton==2.1.0
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.13
wcwidth==0.2.8
Werkzeug==3.0.1
xformers==0.0.22.post7
xxhash==3.4.1
yarl==1.9.2
zipp==3.17.0
